Improve item count logging via htsrv callback.
Instead of assuming the first item to be "viewed" only, we attach a
jquery-appear handler to all of them, and track views asynchronously via
htsrv/log.php.
This makes the view counts much more realistic, because requests from
"unknown" user agents do not get counted anymore, unless they proceed to
to the request.
The old behavior is retained non non-JS behavior, where an IMG tags is
being used in a NOSCRIPT section; only with the first item on the page
however (as previously). This is still an improvement, because the IMG
still has to be requested separately.

Also, instead of looking at the Request URI of a page only to detect
reloads/re-views, the list of viewed items in saved in the users session
instead (with an expiration of 1 week).
An existing(and re-used) session is required for the htsrv request to be
valid (encoded in the `v` verification token`)

This makes the following globals/methods obsolete:
 - $shutdown_count_item_views
 - Hit::is_new_view