Remove rsc_pack=inline hack; add inline js in footer.
Do not create bundles with rsc_pack=inline attributes, but add inline JS
at the end of the page via the new add_footerline function, used by
add_js_headline now.

This allows for bundling jQuery with the regular bundle, too, because
any inline JS is now added to the end of the page.